refactor(cardano-services): hoist metadata queries to a new MetadataService
