+++ packages/cardano-services/src/ChainHistory/DbSyncChainHistory/queries.ts
@@ -179,16 +179,6 @@ export const findRedeemer = `
 	WHERE tx.hash = ANY($1)
 	ORDER BY redeemer.id ASC`;
 
-export const findTxMetadata = `
-	SELECT 
-		meta."key" AS "key",
-		meta."json" AS json_value,
-		tx.hash AS tx_id
-	FROM tx_metadata AS meta
-	JOIN tx ON meta.tx_id = tx.id
-	WHERE tx.hash = ANY($1)
-	ORDER BY meta.id ASC`;
-
 export const findPoolRetireCerts = `
 	SELECT 
 		cert.cert_index AS cert_index,
