+++ packages/cardano-services/test/ChainHistory/DbSyncChainHistoryProvider/ChainHistoryBuilder.test.ts
@@ -162,27 +162,6 @@ describe('ChainHistoryBuilder', () => {
     });
   });
 
-  describe('queryTxMetadataByHashes', () => {
-    test('query transaction metadata by tx hashes', async () => {
-      const result = await builder.queryTxMetadataByHashes([
-        Cardano.TransactionId('3d2278e9cef71c79720a11bc3e08acbbd5f2175f7015d358c867fc9b419ae0b2'),
-        Cardano.TransactionId('545c4656544054045f5a4db0e962f6b09fc6d98b0303d42f3f006e3d920d3720')
-      ]);
-      expect(result.size).toEqual(2);
-      expect(result).toMatchSnapshot();
-    });
-    test('query transaction metadata with empty array', async () => {
-      const result = await builder.queryTxMetadataByHashes([]);
-      expect(result.size).toEqual(0);
-    });
-    test('query transaction metadata when tx not found or has no metadata', async () => {
-      const result = await builder.queryTxMetadataByHashes([
-        Cardano.TransactionId('0000000000000000000000000000000000000000000000000000000000000000')
-      ]);
-      expect(result.size).toEqual(0);
-    });
-  });
-
   describe('queryProtocolParams', () => {
     test('query protocol parameters from last epoch', async () => {
       const result = await builder.queryProtocolParams();
